+++ lib/langchain/llm/aws_bedrock.rb
@@ -11,8 +11,9 @@ module Langchain::LLM
   #
   class AwsBedrock < Base
     DEFAULTS = {
+      chat_completion_model_name: "anthropic.claude-v2",
       completion_model_name: "anthropic.claude-v2",
-      embedding_model_name: "amazon.titan-embed-text-v1",
+      embeddings_model_name: "amazon.titan-embed-text-v1",
       max_tokens_to_sample: 300,
       temperature: 1,
       top_k: 250,
@@ -52,13 +53,11 @@ module Langchain::LLM
     SUPPORTED_CHAT_COMPLETION_PROVIDERS = %i[anthropic].freeze
     SUPPORTED_EMBEDDING_PROVIDERS = %i[amazon cohere].freeze
 
-    def initialize(completion_model: DEFAULTS[:completion_model_name], embedding_model: DEFAULTS[:embedding_model_name], aws_client_options: {}, default_options: {})
+    def initialize(aws_client_options: {}, default_options: {})
       depends_on "aws-sdk-bedrockruntime", req: "aws-sdk-bedrockruntime"
 
       @client = ::Aws::BedrockRuntime::Client.new(**aws_client_options)
       @defaults = DEFAULTS.merge(default_options)
-        .merge(completion_model_name: completion_model)
-        .merge(embedding_model_name: embedding_model)
 
       chat_parameters.update(
         model: {default: @defaults[:chat_completion_model_name]},
@@ -85,7 +84,7 @@ module Langchain::LLM
       parameters = compose_embedding_parameters params.merge(text:)
 
       response = client.invoke_model({
-        model_id: @defaults[:embedding_model_name],
+        model_id: @defaults[:embeddings_model_name],
         body: parameters.to_json,
         content_type: "application/json",
         accept: "application/json"
@@ -180,7 +179,7 @@ module Langchain::LLM
     end
 
     def embedding_provider
-      @defaults[:embedding_model_name].split(".").first.to_sym
+      @defaults[:embeddings_model_name].split(".").first.to_sym
     end
 
     def wrap_prompt(prompt)